+++ lighthouse-core/gather/gather-runner.js
@@ -29,16 +29,6 @@ const UIStrings = {
    */
   warningTimeout: 'The page loaded too slowly to finish within the time limit. ' +
   'Results may be incomplete.',
-  /**
-   * @description Warning that important data was not cleared but may have affected the scores of this run.
-   * @example {IndexedDB, Local Storage} locations
-   */
-  warningData: `{locationCount, plural,
-    =1 {There may be important data in this location: {locations}. ` +
-      `Audit this page in an incognito window to prevent the resources from affecting your scores.}
-    other {There may be important data in these locations: {locations}. ` +
-      `Audit this page in an incognito window to prevent the resources from affecting your scores.}
-  }`,
 };
 
 const str_ = i18n.createMessageInstanceIdFn(__filename, UIStrings);
@@ -131,13 +121,7 @@ class GatherRunner {
     await driver.dismissJavaScriptDialogs();
     await driver.registerRequestIdleCallbackWrap(options.settings);
     if (resetStorage) {
-      const locations = await driver.getImportantLocationsNotCleared(options.requestedUrl);
-      if (locations.length) {
-        LighthouseRunWarnings.push(str_(
-          UIStrings.warningData,
-          {locations: locations.join(', '), locationCount: locations.length}
-        ));
-      }
+      await driver.getImportantStorageWarning(options.requestedUrl, LighthouseRunWarnings);
       await driver.clearDataForOrigin(options.requestedUrl);
     }
     log.timeEnd(status);
