+++ lighthouse-core/test/gather/gather-runner-test.js
@@ -111,6 +111,9 @@ class EmulationDriver extends Driver {
   registerRequestIdleCallbackWrap() {
     return Promise.resolve();
   }
+  getImportantStorageWarning() {
+    return Promise.resolve();
+  }
 }
 
 const fakeDriver = require('./fake-driver.js');
@@ -132,8 +135,7 @@ function resetDefaultMockResponses() {
     .mockResponse('Network.setExtraHTTPHeaders')
     .mockResponse('Network.setUserAgentOverride')
     .mockResponse('Page.enable')
-    .mockResponse('ServiceWorker.enable')
-    .mockResponse('Storage.getUsageAndQuota', {usageBreakdown: []});
+    .mockResponse('ServiceWorker.enable');
 }
 
 beforeEach(() => {
@@ -416,47 +418,13 @@ describe('GatherRunner', function() {
       clearDataForOrigin: createCheck('calledClearStorage'),
       blockUrlPatterns: asyncFunc,
       setExtraHTTPHeaders: asyncFunc,
-      getImportantLocationsNotCleared: () => Promise.resolve(''),
+      getImportantStorageWarning: asyncFunc,
     };
 
     return GatherRunner.setupDriver(driver, {settings: {}}).then(_ => {
       assert.equal(tests.calledCleanBrowserCaches, false);
       assert.equal(tests.calledClearStorage, true);
     });
-  });
-
-  it('warns if important data not cleared may impact performance', async () => {
-    const asyncFunc = () => Promise.resolve();
-    driver.assertNoSameOriginServiceWorkerClients = asyncFunc;
-    driver.beginEmulation = asyncFunc;
-    driver.enableRuntimeEvents = asyncFunc;
-    driver.enableAsyncStacks = asyncFunc;
-    driver.cacheNatives = asyncFunc;
-    driver.registerPerformanceObserver = asyncFunc;
-    driver.dismissJavaScriptDialogs = asyncFunc;
-    driver.registerRequestIdleCallbackWrap = asyncFunc;
-    driver.clearDataForOrigin = asyncFunc;
-
-    connectionStub.sendCommand = createMockSendCommandFn()
-      .mockResponse('Storage.getUsageAndQuota', {usageBreakdown: [
-        {storageType: 'local_storage', usage: 5},
-        {storageType: 'indexeddb', usage: 5},
-        {storageType: 'websql', usage: 0},
-        {storageType: 'appcache', usage: 5},
-        {storageType: 'cookies', usage: 5},
-        {storageType: 'file_systems', usage: 5},
-        {storageType: 'shader_cache', usage: 5},
-        {storageType: 'service_workers', usage: 5},
-        {storageType: 'cache_storage', usage: 0},
-      ]});
-    /** @type {(string | LH.IcuMessage)[]} */
-    const LighthouseRunWarnings = [];
-    await GatherRunner.setupDriver(driver, {settings: {}}, LighthouseRunWarnings);
-    expect(LighthouseRunWarnings[0]).toBeDisplayString(new RegExp(
-      'There may be important data in these locations: Local Storage, IndexedDB. ' +
-      'Audit this page in an incognito window to prevent ' +
-      'the resources from affecting your scores.'
-    ));
   });
 
   it('clears the disk & memory cache on a perf run', async () => {
